Review comment v1.0

+++ runtime/src/iree-amd-aie/aie_runtime/iree_aie_runtime.h
@@ -557,7 +557,7 @@ struct AMDAIEDeviceModel {
     return deviceConfig.preferredLoadBytes;
   }
 
-  uint64_t getMaxRepeatCount(AMDAIETileType tileType) const {
+  uint32_t getMaxRepeatCount(AMDAIETileType tileType) const {
     return devInst.DevProp.DevMod[static_cast<uint8_t>(tileType)]
         .DmaMod->ChProp->MaxRepeatCount;
   }
